Add subscription to movement commands via MQTT
The tank subscribes to commands sent via the MQTT broker and handles 8 different
cases related to changing angle and speed, plus one additional case that stops the
tank and sets its steering angle back to 0. The serial version to control the tank
won't be removed for now for debugging reasons.

Related issue: #6

Co-authored-by: Emma <guslitvem@student.gu.se>